orchestrator/requirements.txt
docker
requests
Django==2.2a1
djangorestframework
django-rest-auth
